+++ lib/auto_shimmer_animate.dart
@@ -1,10 +1,15 @@
-export 'package:shimmer_animation/shimmer_animation.dart'
-    show Shimmer, ShimmerDirection;
-
 export 'src/config/auto_shimmer_config.dart' show AutoShimmerConfig;
 export 'src/config/auto_shimmer_theme.dart' show AutoShimmerTheme;
 export 'src/core/enums/auto_shimmer_direction.dart' show AutoShimmerDirection;
 export 'src/core/typedefs/auto_shimmer_builder.dart' show AutoShimmerBuilder;
+export 'src/effects/auto_shimmer_effect.dart'
+    show
+        AutoShimmerAuroraEffect,
+        AutoShimmerEffect,
+        AutoShimmerPulseEffect,
+        AutoShimmerRawEffect,
+        AutoShimmerSweepEffect;
 export 'src/widgets/auto_shimmer_animate.dart' show AutoShimmerAnimate;
 export 'src/widgets/auto_shimmer_state_animate.dart'
     show AutoShimmerStateAnimate;
+export 'src/widgets/shimmer_wrapper.dart' show AutoShimmerLayer, ShimmerWrapper;